+++ src/main/java/org/c4marathon/assignment/domain/account/service/AccountService.java
@@ -7,17 +7,17 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Random;
 
-import org.c4marathon.assignment.domain.account.dto.response.CreateResponseDto;
 import org.c4marathon.assignment.domain.account.dto.request.RemittanceRequestDto;
-import org.c4marathon.assignment.domain.account.dto.response.RemittanceResponseDto;
 import org.c4marathon.assignment.domain.account.dto.request.SavingRequestDto;
-import org.c4marathon.assignment.domain.account.entity.Account;
-import org.c4marathon.assignment.domain.account.entity.AccountErrCode;
-import org.c4marathon.assignment.domain.account.entity.AccountRole;
-import org.c4marathon.assignment.domain.account.entity.AccountStatus;
-import org.c4marathon.assignment.domain.account.entity.CreateResponseMsg;
-import org.c4marathon.assignment.domain.account.entity.RemittanceResponseMsg;
+import org.c4marathon.assignment.domain.account.dto.response.CreateResponseDto;
+import org.c4marathon.assignment.domain.account.dto.response.RemittanceResponseDto;
 import org.c4marathon.assignment.domain.account.entity.ScheduleCreateEvent;
+import org.c4marathon.assignment.domain.account.entity.account.Account;
+import org.c4marathon.assignment.domain.account.entity.account.AccountRole;
+import org.c4marathon.assignment.domain.account.entity.account.AccountStatus;
+import org.c4marathon.assignment.domain.account.entity.responseMsg.AccountErrCode;
+import org.c4marathon.assignment.domain.account.entity.responseMsg.CreateResponseMsg;
+import org.c4marathon.assignment.domain.account.entity.responseMsg.RemittanceResponseMsg;
 import org.c4marathon.assignment.domain.account.exception.AccountException;
 import org.c4marathon.assignment.domain.account.repository.AccountRepository;
 import org.c4marathon.assignment.domain.account.transaction.TransactionHandler;
